test(runtime): isolate task supervision regressions
